move calendar filter button into the calendar card per mui standard
